api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Sayantan Samanta <sssamanta789@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.3"
flask-restx = "^1.3.0"
google-apps-meet = "^0.1.6"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.134.0"
google-auth-oauthlib = "^1.2.0"
resend = "^2.1.0"
python-dotenv = "^1.0.1"
pika = "^1.3.2"
pydantic = {extras = ["email"], version = "^2.7.4"}
gunicorn = "^22.0.0"
peewee = "^3.17.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
